float...

+++ src/sensor.c
@@ -47,7 +47,7 @@ int initI2c(){
 }
 //https://github.com/emersion/node-i2c-mpu6050を参考にした
 
-#define READ_SB(i) (float)((((int16_t)i2cSensorBuffer[i]) << 8) | i2cSensorBuffer[i+1])
+#define READ_SB(i) ((((int16_t)i2cSensorBuffer[i]) << 8) | i2cSensorBuffer[i+1])
 
 void readSensor(float* ret){
   i2cReadI2CBlockData(i2cHandle,ACCEL_XOUT,i2cSensorBuffer,14);
